Read metadata from separate file
The first idea was to keep non-configuration instance metadata in one
place, i.e., the docker-compose.yml file.  It turns out that
docker-compose does not approve of unknown YAML branches to be included
in the file, so the metadata had to be commented out.  While this
worked, it seemed awkward.

This patch requires users to create metadata.txt files themselves.
Their content is displayed for each instance if --verbose is set; the
very first line is displayed even without --verbose.  Comments (#) in
metadata.txt will be ignored; aside from that, the content is displayed
in verbatim.