entrypoint: Avoid unnecessary warning
The entrypoint should only care if the admin-password-set marker exists
in its volume.  The secrets file is only relevant at the first start.
After that, it's mainly used for reference by osinstancectl.

The file can get deleted without impairing the admin account's security,
so don't mislead with a warning in this case.