Add clone functionality (WIP)
This is a first pass at a simple clone function.  This adds a new
instance with the same OpenSlides server (git) version as the specified
source instance.  Afterwards, it imports the database from the source
instance.

The intended purpose is to create short-lived instances for tests.

What's missing as of now is a 1-to-1 copy of docker-compose.yml
customizations.  We should probably simply copy the file and then update
the port.