Clone: Base docker-compose.yml on source instance's
Before, the configurations of cloned instances were based on the default
docker-compose.yml.example with only a few customizations cherry-picked
from the source instances' configurations.

With this change, source instances' configurations are directly used the
template for the new configurations.  This means that all
customizations, e.g. server commands or mail hostnames, will be copied
from the source instance.

The OpenSlides repository and revision can still be overridden with
command line options.