osinstancectl: Introduce a yellow status indicator
This status signifies instances that are supposed to be running but that
aren't fully functional.

It is to be expected while an instance is starting up but can also
indicate real problems.

Instances in this "unknown" state currently get treated as offline in
terms of filtering (--online/--offline).  Arguments could be made that
they should appear as --online instead or even as both.