server: Include built files in image
So far, volumes were populated during the image build process.  This
meant that to spin up a new OpenSlides instance a new image had to be
built.  Thanks to Docker's caching this was fairly fast but it makes no
sense nonetheless.

More importantly, it prohibits use of other deployment methods which do
not involve an image build process like, e.g., Swarm.

With this patch, all files are built into the image and then copied to
their respective volumes as part of the container's entrypoint script.