Stack: Use NFS as back end
The NFS server is assumed to be available outside the swarm.

Using NFS, both media files and static files can be shared within the
cluster, allowing us to also scale the client freely across different
nodes.

NFS removes the need for the SFTP server which was only used for media
files.  It also limits the extra complexity to the stack configuration,
so for docker-compose setups we can continue to use simple bind mounts.