osinstancectl: Improve parsing of reverse proxy port
This makes sure the script finds the reverse proxy's published port even
if additional ports have been added to the configuration file by hand.

A shortcoming of all of this custom YAML parsing is, inter alia, that it
expects short syntax port mapping definitions.  This is something to
keep in mind while maintaining the template.