osinstancectl: strip custom published ports on clone
When cloning an instance, osinstancectl copies the configuration file
and modifies the reverse proxy's published port to an unused port, so
the instance can be started right away.

It does not modify other ports that users may have added to the file.
In order to prevent errors, osinstancectl will now simply strip custom
port mappings from the configuration and only manage the reverse proxy
port mapping.