osinstancectl: Catch colon in Docker image name
Providing, e.g., openslides:latest as --image would be understandable,
so we'll catch this particular case and helpfully point out --tag.

It would be convenient to simply split such a string and carry on but
then again that doesn't seem worth it.  Most of the time users are
expected to exclusively use --tag for updates; the image name should
change rather infrequently.