entrypoint: Only create settings on prioserver
Before, the entrypoint was explicitly changed to always create the
settings file independently on all server and prioserver instances.  The
reason was that this would allow the server to run on different hosts
without needing to shared a filesystem.  At the time, the SFTP back end
was intended for mediafiles only.

Following the switch to the NFS back end, this is no longer desired.
The prioserver should exclusively be responsible for preparing the
shared personaldata volume.