server: configure REDIS_READ_ONLY_ADDRESS
With this patch, the server entrypoint configures an additional Redis
address in settings.py.  The service (rediscache-slave) is a read-only
replication of the main rediscache service and was added to
docker-compose.yml.example a little while ago.

Once we feel confident that this works as expected we should be able to
remove the redis/rediscache separation.

The REDIS_READ_ONLY_ADDRESS requires a compatible version of the
OpenSlides backend.  In case of problems make sure to rebuild the server
component with an up-to-date version of OpenSlides.