db: Store settings.py in database
The prioserver container generates the Django settings.py file the first
time it is started.  So far, other containers, both other prioservers as
well as regular servers, would access the same file through a volume.

This is harder to achieve with Docker Swarm than with docker-compose
because it requires a network filesystem.  So far we used NFS which has
not been integrated well and becomes a risk as a single point of
failure.

To avoid adding in another complicated component such as a fault
tolerant, distributed filesystem, we can build upon the existing
high-availability database setup.

This patch updates the server entrypoint to store the generated
settings.py in the database.