Add script to update settings.py in database
The settings.py is no longer a single shared file that can be edited on
the host system.  Instead, the version of settings.py that gets put into
production is the one checked into the database.

This patch adds the script update-settings to facilitate the database
update.

The workflow for editing and updating settings.py is now:

 1. enter a server container
 2. edit the settings.py file
 3. run `update-settings` in the server container
 4. exit container and update services