server: Extend config script for arbitrary configs
Besides settings.py additional config files might need to be managed in
the database.  This patch provides the `openslides-config` for this
purpose.

The script still pushes base64 encoded data to the database which is
decoded there.  This has the benefit of easily viewable file contents in
the database; however, it risks corrupting binary data should it ever be
added to the database this way.