osinstancectl: Introduce a Docker Swarm/stack mode
This patch should make most osinstancectl functionality available to
Docker Swarm deployments.

osinstancectl's mode is decided by its invocation: osinstancectl will
continue to work as a docker-compose manager; however, invoked as
osstackctl it will expect a Docker Swarm environment.

Obviously, this is a big change and Swarm is considerably different, so
osstackctl should try to be careful.  For example, it merely suggests
volumes to delete when an instance is removed.

Furthermore, it is not planned to manage other Swarm nodes, e.g., to
clean up persistent volumes.  Currently, there is no need for this
anyway but the slave databases may later be configured with persistent
volumes.