db: force-register standby servers
For now, the standby services do not use a persistent volume.  Instead,
they are set up anew with every start.  For this reason, they need to
register themselves as new services overriding their previous
registrations under the same ID.

In the future, the standby servers should probably also be run with
persistent volumes.