Switch to pgbouncer
This patch replaces the pgpool-II service with a pgbouncer service.

The purpose of pgpool, in this setup, was to be the entry point to the
repmgr cluster.  It was supposed to determine which of the repmgr nodes
is the current primary and treat it accordingly.

This seemed to work at first but pgpool would not reliably find new
primary nodes after failovers.  Furthermore, there can be situations in
which pgpool could not make safe decisions on its own anyway.  Finally,
it turns out that pgpool and repmgr are even considered to be
incompatible.

Therefore, the way forward is to either make pgpool the manager of the
cluster, replacing repmgr, or to keep repmgr and build a service
compatible with it.  The latter is what is being attempted here.

Instead of the built-in failover detection feature of pgpool, the
pgbouncer service leverages Docker's HEALTHECHECK instruction.  Within
the healtcheck script, pgbouncer inquires repmgr about the cluster's
status.  It can find the current primary node and may even make
a decision when there are multiple primaries competing with each other.

The service is marked as unhealthy but may continue to function if the
primary can not be determined with certainty.  It may be worth
considering to completely shut down access to the database to avoid the
risk of writing to the wrong node.

This HEALTHECHECK method is not necessarily limited to pgbouncer but
could also be adapted for pgpool.  However, while the more complex
features of pgpool are not needed it is reasonable to switch to
a simpler tool that also fulfills the requirements.