repmgr: More flexible primary/standby configuration
This change requires an explicit configuration of the primary server in
through an env variable.  It will hopefully make it easier to recreate,
e.g., pgnode1 as a standby of pgnode2.

There are still a lot of assumptions regarding hostnames in the system
that could be addressed at some point.