osinstancectl: Somewhat fix clone function
This updates the clone function for the switch away from pgpool2 and
fixes it for compose mode.

The clone function now speficially clones from pgnode1.  It simply makes
the assumption that pgnode1 is a healthy primary node.

Obviously, this needs to be improved.  An incremental improvement would
be to check the node's status before proceeding, but, ideally, the
connection would be routed through pgbouncer.