repmgr: Split ENTRYPOINT and CMD
This still starts the Postgres cluster using pg_ctlcluster as part of
the entrypoint script.  It might make sense to differentiate the
different aspects of this service further or to introduce options.  For
example, it could come in handy to be able to start Postgres in the
foreground for debugging and backup recovery purposes.