repmgr: Allow backup to fail in promote script
A failure of the basebackup which is created during the standby
promotion process should not indicate a failure of the promotion itself.
Unless the repmgr command itself exits with an error, the script run
should be considered successful.

Testing shows that this change is not actually necessary; repmgr
correctly identifies a successful promotion even when the script exits
with a non-zero code.  However, this patch should make it at least more
clear what is important in the script, and the current behavior may as
well change.

There appears to be no point in adding verbose output to the script
because it will not be displayed by repmgr calling the script.