osinstancectl: Keep first line of metadata empty
The first line of the instances' metadata.txt files is displayed in the
tabular output of "ls" and is supposed to help with the organization of
instances.  The log-like lines that get added automatically by
osinstancectl, e.g., "Instance created on...", clutter the overview
without adding much value, so this patch causes them to start in line 2.
This way, ls has to be invoked with the --metadata option to display the
full metadata.