Do not pre-generate Django secret
Since Django's settings.py gets stored in the database and distributed
to all server nodes this way, there is no longer a need to pre-generate
the Django secret and distribute it as a Docker secret.  We can just let
Django generate its own random key which is a much more common setup.