Store OpenSlides config files in BYTEA
It was assumed that only plain-text configuration files needed to be
stored in the database.  It was an added benefit to be able to view the
file contents plainly in the database.

Even without having run into problems with this setup as of yet the
assumption is risky.  The OpenSlides architecture is under heavy
development and plugins could potentially introduce differing
requirements.

In an attempt to future-proof the setup, all configuration files will be
stored in a BYTEA column

The provided configuration insert script uses a very simple INSERT
method and is, therefore, still limited to very small files.  For larger
files, the script would need to be improved but the schema should be
already prepared for it.