osinstancectl: Configurable instance config template
By default, to create an instance the template repository is cloned.
Afterwards, the included configuration file, e.g.,
docker-stack.yml.example, is used to create the instances configuration.

This patch makes the template configuration file configurable via
/etc/osinstancectl.  This will hopefully reduce the number of manual
edits required for each new instance.

For now, only a single template can be configured which kind of breaks
the ability of docker-composed and Docker Swarm setups to coexist.