osinstancectl: Match instance precisely when using shortcut
In general, osinstancectl takes a pattern as its input that is matched
against the available instance names which may yield multiple results.

Additionally, a shortcut pattern (".") exists that translates the
current working directory into a search pattern.  This is useful for
queries concerning "this instance".

Since the shortcut is supposed to match one specific instance, it is
incorrect to treat the directory name as an open search pattern which
could still potentially yield multiple results.  This patch translates
the name to a precise regexp, ensuring that only the exact instance is
matched by the shortcut.

So far, the precision is only enabled for the ls command.