osinstancectl: vicfg: Suggest other hosts
vicfg needs to be able to connect to a sever containers using
docker-exec which only works on localhost.  In a swarm deployment,
however, the required services may be unavailable on localhost because
they run on other nodes.

With this patch, vicfg's error output becomes a little more helpful.  It
will check on which alternative nodes server services are running and
suggest SSH commands to run vicfg on them.