osinstancectl: No git clone for stacks
For instances deployed as stacks, cloning the full repository for each
instance is unnecessary.  While Compose needs access to the full file
tree because it may attempt to build images on the fly, for Swarm, the
images have to be pre-built and pushed before a deployment anyway.

The intended osinstancectl setup is based on a central template
repository on a local filesystem which is used to, inter alia, build all
required images.  Given this setup, it suffices to copy the few required
files that make up a stack to the instance directory.

In case osstackctl is used without a local template repository, it
should revert to cloning directly; however, this is pretty much
untested.  For Docker Compose, i.e., osinstancectl, the git clone method
remains unchanged either way.