Move from example configurations to templates
Instead of creating YAML configurations from the provided examples once
and editing them, this method relies on settings in .env to persist
customizations.  For changes, the YAML configuration is recreated from
a template.

Of course, the YAML configuration is infinitely customizable, so the
list of implemented configuration variables may always be limiting;
however, the most common per-instance settings should already be
covered.  More fundamental changes to the setup, such as command
overrides or additional services, can be achieved by customizing the
template directly.

Some settings may not need to be added to the YAML file at all, so the
template can probably be simplified later.  Instead of configuring the
services' environments through the YAML configuration, the variables
could be read from .env directly using env_file.

Using a macro language presents the opportunity to merge the Compose and
Swarm templates into a single template file; however, at least for now,
it is more important to avoid the added complexity and maintain two
files.