osinstancectl: Manage configuration .env
From osinstancectl's point of view, docker-compose.yml and
docker-stack.yml have become write-only files.  Configuration options
are written to and read from the projects' .env files.

It is possible for the states of the YAML configuration and .env to get
out of sync; however, the same is true for the YAML configuration
itself.  The only way to be sure of a service's state is to query docker
directly.  It is assumed that instances created with osinstancectl will
always be managed by osinstancectl, so this should be acceptable.