osinstancectl: Remove --default-repo and --mailserver
These were options to the add command, and the settings they affected
are still important; however, there are more settings that may need to
be changed for a fully functional instance.

Even though we could introduce more options and flags, of course, it
becomes less and less likely that they would get used correctly
interactively.  Instead, some customizations should simply be made in
.env or, in case of system-global settings such as the default Docker
repository, in a customized template.

Custom template paths can be set in /etc/osinstancectl.  Adding command
line options for template paths is something we could definitely add if
there is demand for it.