osinstancectl: Re-deploy for updates
For stacks, osinstancectl used to run `docker service update` precisely
for the services affected by the update.  For this command, the complete
"image:tag" name has to be known which is possible if we parse it from
the YAML configuration.  The .env-based configuration makes image names
optional, however, so osinstancectl can not acquire all required
information from the .env file alone.

The solution chosen here is to regenerate the YAML configuration file in
order to pick up, for example, updated tags, and to re-deploy the whole
stack.

This means trusting that Swarm will not unnecessarily disrupt services
that should be unaffected.  Should this approach fail, we could either
require all image information to be present in .env, retrieve the
missing information from `docker service inspect`, or go back to parsing
it from the YAML configuration.