Config: Make repmgr primary configurable through .env
It should rarely be necessary to manually edit these values which is why
they have not been added to .env at all.

Should there be a regular use case for changing these values the
template and database Docker image should be improved to allow setting
the primary node with a single variable.