Config: Scaling based on .env in docker-compose.yml
The Docker Compose template will now conditinally add deploy.replicas
settings to the server and redis-slave services if the defaults have
been changed in .env.  As with the Docker Stack template, additional
services could be made configurable as needed.

The replica settings are non-standard, and for docker-compose to pick up
these settings, it must be invoked with --compatibility.