Config: Enable secrets based on file availability
If m4 can find secrets/adminsecret.env and/or secrets/usersecret.env it
will include the Docker secrets configuration which was the default
until now.

This change is in preparation of future changes to the instance setup
and will, hopefully, simplify its setup instructions.

It might also be a good idea to merge the secrets files.  They are both
intended for the same service and their contents do not conflict with
each other.