repmgr: Automatically rejoin former primaries
During database failover events, repmgr elects a new primary node and
pgbouncer follows that decision.  If the original primary node comes
back online at a later point in time, it can not simply rejoin the
replication cluster as another primary.

Until now, the solution was to delete the former primary cluster and
recreate it as a new standby.  The procedure is simple but requires
manual intervention.  With this patch, former primaries should be able
to automatically rejoin their replication clusters as standby nodes.

The process is roughly as follows:

A starting primary Postgres/repmgr node will check which node is
currently considered the replication clusters' primary node by
pgbouncer.  If pgbouncer considers another node to be primary, the new
node will double-check the node's primary status by contacting it
directly.  If the other node has indeed become the primary, the starting
node will relinquish its primary status to become a standby node.