repmgr: Use regular pg_ctlcluster during setup
Postgres is generally started by supervisorctl, so its status and
logging work as expected.

During the cluster setup procedure, however, repmgr may have to start
the cluster temporarily.  Since the setup occurs before even supervisord
is started, the repmgr.conf file is changed to set the startup command
to a regular pg_ctlcluster command.  After the initial setup, the
configuration is changed to a supervisorctl command.