osinstancectl: Add mid-erase hook
This awkwardly named hook is run during the deletion of Swarm instances.
Instances get stopped first; afterwards, their project directory is
removed.  The hook is executed between these two steps.

Its intended application is deleting the instance's volumes.  To delete
volumes they must not be in use, and to identify them the project
directory must not have been removed yet.

Volume deletion is implemented as a hook for now because the actual
deletion process may vary from swarm to swarm and require customization.

On the other hand, the recently introduced vicfg changes require SSH
access to other nodes already which is one way to handle volume deletion
on other nodes.  It should be acceptable to add a built-in SSH-based
method for volume deletion as long it is an opt-in option.  Either way,
a way to customize a delicate process such as data removal will still be
desirable then.

This patch also removes the post-remove and post-erase hooks because
they cannot actually be executed, at least with the current
implementation of run_hook().