Contrib: Remove legacy-import script
legacy-import.sh is incompatible with current setups.  The most
significant change in this context is that the settings.py is no longer
stored in the database (instancecfg); instead, .env is used for
configuration.

As well as modifying the legacy /export/ script, an updated legacy
import script would need to parse the old settings.py file to generate
an equivalent .env file.

The openslides and mediafile databases could be populated the same way
as before, so the previous migration concept would still be a good basis
for further development.