osinstancectl: Update without redeploying entire stack
In commit 7ca785d75491501607b5bea2512d23859c76187a the update procedure
for stacks was simplified.  An update would simply rewrite the
docker-stack.yml configuration and then redeploy the stack.  In most
cases, this has worked well; however, in some cases unrelated services
of the stack would get restarted and/or moved to other nodes during the
deployment step.  This has undesired consequences when the Redis
services are affected by this because users may lose their session.

This patch reverts the update procedure to update the server and client
images specifically, ensuring that the rest of the stack is left alone.