pgbouncer: Improve HEALTHCHECK logic
This patch causes PgBouncer operations to be paused when a primary
database cannot be found and to exit cleanly afterwards.  This, in turn,
has the effect that HEALTHCHECK will not consider a PgBouncer container
unhealthy when it cannot make a connection to a primary database.  In
this way, instead of killing unhealthy containers, HEALTHCHECK merely
takes care of continually running the HEALTHCHECK script which will keep
retrying to find a primary.

Since the PgBouncer container does not get killed by HEALTHCHECK, this
also enables users to manually pause PgBouncer indefinitely while taking
the database offline.  Before, PgBouncer operations would surprisingly
resume when a new container was spawned due to the failing HEALTHCHECK.

The HEALTHCHECK timeout setting has been raised considerably in order to
not conflict with the time required for the HEALTHCHECK script to
complete.  The Script uses internal timeouts.  The HEALTHCHECK timeout
should only take effect if an unforeseen error condition causes the
script to get stuck.