Make searches case-insenstive
This is actually a work-around for the fact that we use the given
argument string as both a search term and the PROJECT_NAME.  At one
point we had to force PROJECT_NAMEs to lower-case to avoid various
problems, making it impossible to properly use the same string for
regular expressions in the process.

This patch makes all relevant regular expression searches
case-insensitive which restores a predictable search function.  However,
a better solution would differentiate between search terms and
(internal) project names.