osinstancectl: Rename proxy options and variables
HAProxy was replaced by Caddy in the OpenSlides stack.  At the same
time, various variables and options were renamed to the generic term
"proxy".  This patch mirrors this change for osinstancectl, both
regarding its parameters as well as internally, so future changes of the
reverse proxy component should not require another patch.